21.11.12
1. 모든 예상되는 오류에 대해 모달창으로 알리는 작업 완료
2. Service 클래스에서 Exception을 throw하도록 재작성, Service에서 사용하지 않는 메소드 삭제
3. DnF/LoL Controller의 메소드 순서 조정

+++ dnflol/src/main/java/com/project/dnflol/Service/DApplyService.java
@@ -9,6 +9,7 @@ import com.project.dnflol.DAO.DApplyDAO;
 import com.project.dnflol.DAO.DGroupDAO;
 import com.project.dnflol.DTO.DApplyDTO;
 import com.project.dnflol.Exception.AlreadyExistedApplyException;
+import com.project.dnflol.Exception.NoSuchApplyException;
 import com.project.dnflol.Exception.TooManyApplyException;
 
 @Service
@@ -38,12 +39,15 @@ public class DApplyService {
 		dapplyDao.createToAccepted(dapplyDto);
 	}
 	
-	public List<DApplyDTO> readAllAcceptedByGroupId(int dgroupId) {
-		return dapplyDao.readAllAcceptedByGroupId(dgroupId);
+	public DApplyDTO read(DApplyDTO dapplyDto) {
+		DApplyDTO apply = dapplyDao.read(dapplyDto);
+		if (apply == null)
+			throw new NoSuchApplyException("해당하는 가입 요청이 없습니다.");
+		return apply;
 	}
 	
-	public DApplyDTO read(DApplyDTO dapplyDto) {
-		return dapplyDao.read(dapplyDto);
+	public List<DApplyDTO> readAllAcceptedByGroupId(int dgroupId) {
+		return dapplyDao.readAllAcceptedByGroupId(dgroupId);
 	}
 	
 	public List<DApplyDTO> readAllByGroupId(int dgroupId) {
@@ -62,6 +66,8 @@ public class DApplyService {
 	}
 	
 	public void delete(DApplyDTO dapplyDto) {
-		dapplyDao.delete(dapplyDto);
+		DApplyDTO apply = read(dapplyDto);
+		if (apply != null)
+			dapplyDao.delete(dapplyDto);
 	}
 }
